backend/requirements.txt
PyJWT==1.7.1
sanic==18.12.0
Sanic-Cors==0.10.0.post3